Fixed basic syntax issues.
